fix: fixed the re rendering of sidebar when user switches between videos
