+++ src/app/courses/[courseId]/loading.tsx
