+++ src/components/CourseCard.tsx
@@ -21,12 +21,12 @@ export const CourseCard = ({
       <div className="relative">
         {course.totalVideos !== undefined &&
           course.totalVideosWatched !== undefined && (
-          <PercentageComplete
-            percent={Math.ceil(
-              (course.totalVideosWatched / course.totalVideos) * 100,
-            )}
-          />
-        )}
+            <PercentageComplete
+              percent={Math.ceil(
+                (course.totalVideosWatched / course.totalVideos) * 100,
+              )}
+            />
+          )}
       </div>
       <img src={course.imageUrl} alt={course.title} className="rounded-md" />
       <div className="p-2">
