+++ src/icons/BackArrow.tsx
@@ -1,11 +1,11 @@
-export const BackArrow = () => {
+export const BackArrow = ({ ...props }) => {
   return (
     <svg
-      className="w-5 h-5 rtl:rotate-180"
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
       stroke="black"
+      {...props}
     >
       <path d="M6.75 15.75L3 12m0 0l3.75-3.75M3 12h18" />
     </svg>
